pyproject.toml
[project]
name = "hollerback"
version = "0.1.0"
description = "A Signal Messenger gateway for your local Goose agent"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Theron Conrey", email = "theron@conrey.org" },
]
keywords = ["signal", "goose", "mcp", "gateway", "ai", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: No Input/Output (Daemon)",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Internet",
]
dependencies = [
    "click>=8.3.2",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "pyyaml>=6.0.3",
    "rich>=15.0.0",
]

[project.urls]
Homepage = "https://github.com/theronconrey/hollerback"
Source = "https://github.com/theronconrey/hollerback"
"Bug Tracker" = "https://github.com/theronconrey/hollerback/issues"

[project.scripts]
hollerback = "hollerback.cli:main"

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
